chore: helper component refactor and add id to clickable's toggle item
