+++ packages/box/docs/Box.mdx
@@ -1,4 +1,5 @@
-import { Box, Clickable } from '../src';
+import { Box } from '../src';
+import { Clickable } from '../../_helpers';
 
 # Box
 
@@ -86,6 +87,6 @@ import { Box } from '@fabric-ds/react';
 
 ## Clickable props
 
-```props packages/box/src/clickable.tsx
+```props packages/_helpers/clickable.tsx
 
 ```
