+++ packages/box/src/index.ts
@@ -1,3 +1,2 @@
 export { Box } from './component';
-export { Clickable } from './clickable';
-export type { BoxProps, ClickableProps } from './props';
+export type { BoxProps } from './props';
