+++ packages/box/src/props.tsx
@@ -57,43 +57,3 @@ export interface BoxProps {
    */
   neutral?: boolean;
 }
-
-export type ClickableProps = {
-  /**
-   * Passes radio type to the underlying toggle
-   */
-  radio?: boolean;
-
-  /**
-   * Passes checkbox type to the underlying toggle
-   */
-  checkbox?: boolean;
-
-  /**
-   * Value of the dead toggle
-   */
-  value?: string;
-
-  /**
-   * Clickable element children
-   */
-  children: JSX.Element | JSX.Element[] | string;
-
-  /**
-   * Redirect to url on click
-   * If passed, clickable renders as an anchor tag allowing you to pass properties such as target, rel, etc.
-   */
-  href?: string;
-
-  /**
-   * Additional classnames to the toggle label
-   */
-  labelClassName?: string;
-
-  /**
-   * Click handler
-   * If passed, clickable renders as a button with a click event
-   */
-  onClick?: () => void;
-} & Partial<Omit<HTMLAnchorElement, 'children'>> &
-  Partial<Omit<HTMLButtonElement, 'children'>>;
