+++ packages/card/docs/Card.mdx
@@ -1,5 +1,5 @@
-import { Clickable } from '../../box/src';
-import { DeadToggle, Toggle } from '../../toggle/src';
+import { Toggle } from '../../toggle/src';
+import { Clickable, DeadToggle } from '../../_helpers';
 import { Card } from '../src';
 
 # Card
@@ -456,5 +456,11 @@ function Example() {
 # Props
 
 ```props packages/card/src/component.tsx
+
+```
+
+## DeadToggle props
+
+```props packages/_helpers/dead-toggle.tsx
 
 ```
