+++ packages/toggle/src/index.ts
@@ -1,3 +1,2 @@
 export { Toggle } from './component';
-export { DeadToggle } from './dead-toggle';
 export type { ToggleProps } from './props';
